requirements.txt
asgiref==3.8.1
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-rest-auth==6.0.0
Django>=4.2.16
django-allauth==65.0.2
django-cors-headers==4.4.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
